% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPC.R
\name{CME_select}
\alias{CME_select}
\title{Variable selection with CEM estimator}
\usage{
CME_select(
  Y,
  X,
  ky,
  kx,
  num_features,
  eps,
  appro = F,
  tol = 0.001,
  numCores = 1
)
}
\arguments{
\item{Y}{a matrix of responses (n by dy)}

\item{X}{a matrix of predictors (n by dx)}

\item{ky}{a function \eqn{k(y, y')} of class \code{kernel} on the space of \eqn{Y}. It can be the kernel implemented in \code{kernlab} e.g. \code{rbfdot(sigma = 1)}, \code{vanilladot()}}

\item{kx}{a list of length \code{num_features}; \code{kx[[k]]} is the kernel used for \code{(Xj1,...,Xjk)}, the first \code{k} selected variables.}

\item{num_features}{the number of variables to be selected, cannot be larger than dx.}

\item{eps}{a positive number; the regularization parameter for CME estimator}

\item{appro}{whether to use incomplete Cholesky decomposition for approximation}

\item{tol}{tolerance used for incomplete Cholesky decomposition (\code{inchol} in package \code{kernlab})}

\item{numCores}{number of cores that are going to be used for parallelizing the process.}
}
\value{
a vector of the indices from \code{1,...,dx} of the selected variables
}
\description{
The algorithm performs a forward stepwise variable selection using CME estimators.
}
\details{
A stepwise forward selection of variables using KPC. At each step the \eqn{Xj} maximizing \eqn{\tilde{rho^2}(Y,Xj | selected Xi)} is selected.
}
\examples{
n = 200
p = 100
X = matrix(rnorm(n * p), ncol = p)
Y = X[, 1] * X[, 2] + sin(X[, 1] * X[, 3]) + rnorm(n)
kx = c(kernlab::rbfdot(1),kernlab::rbfdot(1/2),kernlab::rbfdot(1/3))
CME_select(Y, X, rbfdot(1), kx, 3, eps = 1e-3, appro = F, numCores = 1)
}
