% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPC.R
\name{Klin}
\alias{Klin}
\title{\eqn{\hat{\eta}_n^{\mbox{lin}}} with geometric graphs.}
\usage{
Klin(Y, X, k, Knn = 1)
}
\arguments{
\item{Y}{a matrix of response (n by dy)}

\item{X}{a matrix of predictors (n by dx)}

\item{k}{a function \eqn{k(y, y')} of class \code{kernel}. It can be the kernel implemented in \code{kernlab} e.g. \code{rbfdot(sigma = 1)}, \code{vanilladot()}}

\item{Knn}{the number of K-nearest neighbor to use; or "MST".}
}
\value{
klin: an empirical kernel measure of association which can be computed in near linear time when Knn graphs are used.
}
\description{
Calculate \eqn{\hat{\eta}_n^{\mbox{lin}}} (the unconditional version of graph-based KPC) using directed Knn graph or minimum spanning tree (MST).
The computational complexity is O(nlog(n))
}
\details{
\eqn{\hat{\eta}_n} is an estimate of the population kernel measure of association, based on data \eqn{(X_1,Y_1),\ldots ,(X_n,Y_n)\sim \mu}.
For Knn graph, \eqn{\hat{\eta}_n} can be computed in near linear time (in \eqn{n}).
In particular,
\deqn{\hat{\eta}_n^{\mbox{lin}}:=\frac{n^{-1}\sum_{i=1}^n d_i^{-1}\sum_{j:(i,j)\in\mathcal{E}(G_n)} K(Y_i,Y_j)-(n-1)^{-1}\sum_{i=1}^{n-1} K(Y_i,Y_{i+1})}{n^{-1}\sum_{i=1}^n K(Y_i,Y_i)-(n-1)^{-1}\sum_{i=1}^{n-1} K(Y_i,Y_{i+1})}},
where all symbols have their usual meanings as in the definition of \eqn{\hat{\eta}_n}.
}
\examples{
library(kernlab)
Klin(Y = rnorm(100), X = rnorm(100), k = rbfdot(1), Knn = 1)
}
\references{
Deb, N., P. Ghosal, and B. Sen (2020). Measuring association on topological spaces using kernels and geometric graphs
}
