% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPC.R
\name{KPCCME}
\alias{KPCCME}
\title{KPCCME}
\usage{
KPCCME(Y, X = NULL, Z, ky, kx, kxz, eps, tol = 0.001)
}
\arguments{
\item{Y}{Matrix (n by dy)}

\item{X}{Matrix (n by dx) or NULL if X is empty}

\item{Z}{Matrix (n by dz)}

\item{ky}{a kernel function k(y,y')}

\item{kx}{kernel function for X}

\item{kxz}{kernel function for (X,Z) or for Z if X is empty}

\item{eps}{A small regularization parameter for inverting the empirical cross-covariance operator}

\item{tol}{Tolerance used for incomplete Cholesky decomposition}
}
\description{
Calculate KPC rho^2(Y,Z|X) with CME method
}
\details{
kx,ky,kxz should be kernels provided in the R package "kernlab", so that kernelMatrix can be applied
}
\examples{
data("medical")
KPCCME(med$D,med$C,med$U,rbfdot(1/(2*median(dist(med$D))^2)),rbfdot(1/(2*median(dist(med$C))^2)), rbfdot(1/(2*median(dist(cbind(med$C,med$U)))^2)), 1e-3,1e-12) # 0.0003175344 D is independent of U given C
KPCCME(med$D,med$U,med$C,rbfdot(1/(2*median(dist(med$D))^2)),rbfdot(1/(2*median(dist(med$U))^2)), rbfdot(1/(2*median(dist(cbind(med$C,med$U)))^2)), 1e-3,1e-12) # 0.6834605 D is associated with C controlling U
}
