% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPC.R
\name{KPCCMElinear}
\alias{KPCCMElinear}
\title{Kernel partial correlation with CME method using linear kernels}
\usage{
KPCCMElinear(Y, X = NULL, Z, eps)
}
\arguments{
\item{Y}{a matrix (n by dy)}

\item{X}{a matrix (n by dx) or \code{NULL} if \eqn{X} is empty}

\item{Z}{a matrix (n by dz)}

\item{eps}{a small regularization parameter for inverting the empirical cross-covariance operator}
}
\description{
Linear kernels are used for ky, kx, kxz in KPCCME.
}
\details{
Linear kernels are used for ky, kx, kxz in KPCCME.
The kernel partial correlation squared (KPC) measures the conditional dependency
between \eqn{Y} and \eqn{Z} given \eqn{X}, based on an i.i.d. sample of \eqn{(Y, Z, X)}.
It converges to the population quantity which is between 0 and 1.
A small value indicates low conditional dependency between \eqn{Y} and \eqn{Z} given \eqn{X}, and
a large value indicates stronger conditional dependence.
If \code{X = NULL}, it measures the unconditional dependency between \eqn{Y} and \eqn{Z}.
}
\examples{
n = 2000
x = rnorm(n)
z = rnorm(n)
y = x + z + rnorm(n,1,1)
KPCCMElinear(y, x, z, 1e-5/n^(0.4))
}
