% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPC.R
\name{KPCKnn}
\alias{KPCKnn}
\title{KPC (with Knn graph)}
\usage{
KPCKnn(Y, X, Z, k, Knn = 1, trans_inv = FALSE)
}
\arguments{
\item{Y}{Matrix (n by dy)}

\item{X}{Matrix (n by dx)}

\item{Z}{Matrix (n by dz)}

\item{k}{a kernel function k(y,y')}

\item{Knn}{number of nearest neighbor to use}

\item{trans_inv}{TRUE or FALSE. Is kernel k translation invariant?}
}
\description{
Calculate the kernel partial correlation squared (KPC) with directed Knn graph
}
\details{
The kernel partial correlation squared (KPC) measures the conditional dependency
between Y and Z given X, based on an i.i.d. sample of (Y, Z, X).
The coefficient is asymptotically guaranteed to be between 0 and 1.

The value returned by KPC can be positive or negative. Asymptotically, it is guaranteed
  to be between 0 and 1. A small value indicates low conditional dependence between Y and Z given X, and
  a high value indicates strong conditional dependence.
}
\examples{
data("medical")
KPCKnn(med$D,med$C,med$U,kernlab::rbfdot(1/(2*median(dist(med$D))^2)),trans_inv = T)
}
