% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPC.R
\name{KPCRKHSlinear}
\alias{KPCRKHSlinear}
\title{Kernel partial correlation with RKHS method using linear kernels}
\usage{
KPCRKHSlinear(Y, X = NULL, Z, eps = 0.001)
}
\arguments{
\item{Y}{a matrix (n by dy)}

\item{X}{a matrix (n by dx) or \code{NULL} if \eqn{X} is empty}

\item{Z}{a matrix (n by dz)}

\item{eps}{a small regularization parameter for inverting the empirical cross-covariance operator}
}
\value{
The algorithm returns a real number which is the estimated KPC (with linear kernel).
}
\description{
Linear kernels are used for ky, kx, kxz in KPCRKHS, in which case the incomplete Cholesky decomposition is the data matrix itself and could speed up the computation to a great extent.
}
\details{
Linear kernels are used for ky, kx, kxz in KPCRKHS.
The kernel partial correlation (KPC) coefficient measures the conditional dependence
between \eqn{Y} and \eqn{Z} given \eqn{X}, based on an i.i.d. sample of \eqn{(Y, Z, X)}.
It converges to the population quantity which is between 0 and 1.
A small value indicates low conditional dependence between \eqn{Y} and \eqn{Z} given \eqn{X}, and
a large value indicates stronger conditional dependence.
If \code{X = NULL}, it measures the unconditional dependence between \eqn{Y} and \eqn{Z}.
}
\examples{
n = 2000
x = rnorm(n)
z = rnorm(n)
y = x + z + rnorm(n,1,1)
KPCRKHSlinear(y, x, z, 1e-5/n^(0.4))
}
\seealso{
\code{\link{KPCRKHS}}
}
