% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{med}
\alias{med}
\title{Medical data from 35 patients}
\format{
A data frame with 35 rows and 3 variables:
\describe{
  \item{C}{creatinine clearance, in ml/min/1.73m^2}
  \item{D}{digoxin clearance, in ml/min/1.73m^2}
  \item{U}{urine flow, in ml/min}
}
}
\source{
Edwards, D. (2012). Introduction to graphical modelling, Section 3.1.4, Springer Science & Business Media.
}
\usage{
med
}
\description{
A dataset containing three variables (creatinine clearance C; digoxin clearance D; urine flow U) from 35 patients.
}
\keyword{datasets}
