% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPC.R
\name{MSE}
\alias{MSE}
\title{Calculating the mean squared error}
\usage{
MSE(X, Y, kx, ky, eps)
}
\arguments{
\item{X}{a matrix of predictors (n by dx)}

\item{Y}{a matrix of responses (n by dy)}

\item{kx}{the kernel function for \eqn{X}}

\item{ky}{a function \eqn{k(y, y')} of class \code{kernel} on the space of \eqn{Y}. It can be the kernel implemented in \code{kernlab} e.g. \code{rbfdot(sigma = 1)}, \code{vanilladot()}}

\item{eps}{the regularization parameter for the CME estimator}
}
\description{
Return the mean squared error between \eqn{k(Y_i,\cdot)} and the empirical CME \eqn{\hat{\mu}_{Y|Xi}}.
}
\details{
The MSE provides a way to select the regularization parameter \code{eps} in the CME estimator.
\code{eps} should be chosen as small as possible while keeping the MSE reasonably small.
}
\examples{
n = 200
p = 100
X = matrix(rnorm(n * p), ncol = p)
Y = X[, 1] * X[, 2] + sin(X[, 1] * X[, 3]) + rnorm(n)
for (eps in 10^(-(1:9))) print(MSE(X[,1],Y,kernlab::rbfdot(1),kernlab::rbfdot(1),eps))
for (eps in 10^(-(1:9))) print(MSE(X[,1:2],Y,kernlab::rbfdot(1/2),kernlab::rbfdot(1),eps))
for (eps in 10^(-(1:9))) print(MSE(X[,1:3],Y,kernlab::rbfdot(1/3),kernlab::rbfdot(1),eps))
}
