% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ElecData}
\alias{ElecData}
\title{2017 Korea presidential election data}
\format{
A data frame with 1250 rows and 9 variables:
\describe{
  \item{PrecinctCode}{250 precinct codes designated by the election committee (4 digits)}
  \item{CityCode}{250 city codes of administrative standard code management system (5 digits)}
  \item{CandidateName}{Symbols 1-5, corresponding to Moon Jae-in, Hong Jun-pyo, Ahn Cheol-soo, Yoo Seung-min, Shim Sang-jung}
  \item{AveAge}{Average age of voters in 17 years: statistics on resident registration population of the Ministry of Government Administration and Home Affairs}
  \item{AveYearEdu}{Average number of years of education for voters}
  \item{AveHousePrice}{Average price per square meter in 17 years}
  \item{AveInsurance}{The average insurance premium for each city, county, district}
  \item{VoteRate}{Vote rate by candidate}
  \item{NumVote}{Number of votes by candidate}
}
}
\source{
\url{https://github.com/OhmyNews/2017-Election}
}
\usage{
ElecData
}
\description{
A dataset containing 9 variables, consists of the voting results earned by the top five candidates from 250 electoral districts in Korea.
}
\keyword{datasets}
