% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPC.R
\name{RKHS_select}
\alias{RKHS_select}
\title{Variable selection with RKHS estimator}
\usage{
RKHS_select(
  Y,
  X,
  ky,
  kx,
  num_features,
  eps,
  appro = F,
  tol = 0.001,
  numCores = 1,
  verbose = F
)
}
\arguments{
\item{Y}{a matrix of responses (n by dy)}

\item{X}{a matrix of predictors (n by dx)}

\item{ky}{a function \eqn{k(y, y')} of class \code{kernel}. It can be the kernel implemented in \code{kernlab} e.g. Gaussian kernel: \code{rbfdot(sigma = 1)}, linear kernel: \code{vanilladot()}. In practice, Gaussian kernel with empirical bandwidth \code{kernlab::rbfdot(1/(2*median(dist(Y))^2))} may be a good choice.}

\item{kx}{a list of length \code{num_features}; \code{kx[[k]]} is the kernel used for \code{(Xj1,...,Xjk)}, the first \code{k} selected variables.}

\item{num_features}{the number of variables to be selected, cannot be larger than dx.}

\item{eps}{a positive number; the regularization parameter for the RKHS estimator}

\item{appro}{whether to use incomplete Cholesky decomposition for approximation}

\item{tol}{tolerance used for incomplete Cholesky decomposition (\code{inchol} in package \code{kernlab})}

\item{numCores}{number of cores that are going to be used for parallelizing the process.}

\item{verbose}{whether to print each selected variables during the forward stepwise algorithm}
}
\value{
a vector of the indices from \code{1,...,dx} of the selected variables
}
\description{
The algorithm performs a forward stepwise variable selection using RKHS estimators.
}
\details{
A stepwise forward selection of variables using KPC. At each step the \eqn{Xj} maximizing \eqn{\tilde{rho^2}(Y,Xj | selected Xi)} is selected.
}
\examples{
n = 200
p = 100
X = matrix(rnorm(n * p), ncol = p)
Y = X[, 1] * X[, 2] + sin(X[, 1] * X[, 3]) + rnorm(n)*0.5
library(kernlab)
kx = c(rbfdot(1),rbfdot(1/2),rbfdot(1/3))
RKHS_select(Y, X, rbfdot(1), kx, 3, eps = 1e-3, appro = F, numCores = 1)
}
