% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KPC.R
\name{TnKnn}
\alias{TnKnn}
\title{Tn with Knn graph}
\usage{
TnKnn(Y, X, k, Knn = 1)
}
\arguments{
\item{Y}{Matrix of response (n by dy)}

\item{X}{Matrix of predictors (n by dx)}

\item{k}{a kernel function k(y,y')}

\item{Knn}{The number of K-nearest neighbor to use}
}
\value{
Tn: a numeric number
}
\description{
Calculate Tn using directed Knn graph.
}
\details{
Tn is an estimate of E[E[k(Y1,Y1')|X]], with Y1, Y1' drawn iid from Y|X, given X.
}
